% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_controlFilter.R
\name{myFUN_controlFilter}
\alias{myFUN_controlFilter}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_controlFilter(Entry)
}
\arguments{
\item{Entry}{nummer of Population (in myGWC )}
}
\value{
returns a dataframe with the filtered values
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_controlFilter()

}
\keyword{filter}
