% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_droughtFilter.R
\name{myFUN_droughtFilter}
\alias{myFUN_droughtFilter}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_droughtFilter(Entry, cutWeight = 370, cutGWC = 1.5)
}
\arguments{
\item{Entry}{nummer of Population (in myGWC )}

\item{cutWeight}{maximal weight which should be showed, default= 370}

\item{cutGWC}{maximal GWC which should be showed, default= 1.5}
}
\value{
returns a dataframe with the filtered values
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_droughtFilter()

}
\keyword{filter}
