% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_ggplotRegression.R
\name{myFUN_ggplotRegression}
\alias{myFUN_ggplotRegression}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_ggplotRegression(fit, genotype = NULL)
}
\arguments{
\item{fit}{output of a lm regression}

\item{genotype}{genotype, for the title, default = NULL}
}
\value{
ggplot
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_ggplotRegression()

}
\keyword{ggplot}
