% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_ggscatter.R
\name{myFUN_ggscatter}
\alias{myFUN_ggscatter}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_ggscatter(df, xaxis, yaxis, xlab = xaxis, ylab = yaxis,
  regression = "reg.line", xlimit = c(0, 1.6), ylimit = NULL)
}
\arguments{
\item{df}{data.frame which contains the data}

\item{xaxis}{data on the x axis of the plot}

\item{yaxis}{data on the y axis of the plot}

\item{regression}{method of the regression, defaulkt = "reg.line"}

\item{xlimit}{x-axis definiton, default = c(0, 1.6)}

\item{ylimit}{y-axis definiton, default = NULL}

\item{xalab}{title of the xlab, default = xaxis}

\item{yalab}{title of the ylab, default = yaxis}
}
\value{
returns a dataframe with the filtered values
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_ggscatter()

}
\keyword{filter}
