% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_LERsmoother.R
\name{myFUN_LERsmoother}
\alias{myFUN_LERsmoother}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_LERsmoother(entry, smoothFactor = 20, column = "ler")
}
\arguments{
\item{entry}{nummer of Population (in myGWC )}

\item{smoothFactor}{number, the higher the smoother, default =20}

\item{column}{solumn in data.frame which should be smoothed, default = "ler"}
}
\value{
myGWC[[entry]] with smoothed values
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_LERsmoother()

}
\keyword{filter}
