% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_MODELsmoother.R
\name{myFUN_MODELsmoother}
\alias{myFUN_MODELsmoother}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_MODELsmoother(entry, smoothFactor = 10)
}
\arguments{
\item{entry}{nummer of Population (in myGWC )}

\item{smoothFactor}{number, the higher the smoother, default =20}
}
\value{
myGWC[[entry]] with smoothed values
}
\description{
filtering data to drought treatment relevant
}
\examples{
myFUN_MODELsmoother()

}
\keyword{filter}
