% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFUN_myGenotypeParametersControl.R
\name{myFUN_myGenotypeParametersControl}
\alias{myFUN_myGenotypeParametersControl}
\title{filter nfunction --> GLEIS}
\usage{
myFUN_myGenotypeParametersControl(Geno, input1 = "ler_smooth_diff",
  input2 = "gwc_dry")
}
\arguments{
\item{Geno}{single experiment}

\item{input1}{input 1 for the regression}

\item{input2}{input 2 for the regression}
}
\value{
returns a vector with(Population, PlantID, Entry, Treatment, R2, P, Intercept, Slope)
}
\description{
Parametrisieren der Drought-Treatments
}
\examples{
myFUN_myGenotypeParametersControl()
}
\keyword{parametrisieren}
